+ rename test_social module to algorithm_sandbox
