+++ .gitignore
@@ -1,5 +1,5 @@
 # JetBrains IDEA
 /.idea/
 /venv/
-/test_social/mock_data/
-/test_social/logs/
+/algorithm_sandbox/mock_data/
+/algorithm_sandbox/logs/
