+++ algorithm_sandbox/mock_team_generation.py
@@ -6,7 +6,7 @@ from team_formation.app.team_generator.algorithm.consts import FRIEND, ENEMY, DE
 from team_formation.app.team_generator.algorithm.social_algorithm.social_algorithm import SocialAlgorithm
 from team_formation.app.team_generator.student import Student
 from team_formation.app.team_generator.team_generator import TeamGenerationOption, TeamGenerator
-from test_social.encoder import Encoder, load_json_data
+from algorithm_sandbox.encoder import Encoder, load_json_data
 
 
 class MockData:
